TelkomAthon1/requirements.txt
# Core dependencies
streamlit==1.31.0
openai==1.12.0
psycopg2-binary==2.9.9
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1

# Testing dependencies
hypothesis==6.98.0
pytest==8.0.0
pytest-cov==4.1.0
